refactor: rename register action
